change polyfills location, rename support dir
