+++ packages/base/src/sap/ui/webcomponents/base/browsersupport/Edge.js
