+++ packages/main/bundle.esm.js
@@ -1,5 +1,5 @@
 // ESM bundle targets Edge + browsers with native support
-import "@ui5/webcomponents-base/src/sap/ui/webcomponents/base/support/Edge";
+import "@ui5/webcomponents-base/src/sap/ui/webcomponents/base/browsersupport/Edge";
 
 import "@ui5/webcomponents-base/src/sap/ui/webcomponents/base/shims/jquery-shim";
 import "@ui5/webcomponents-base/src/sap/ui/webcomponents/base/events/PolymerGestures";
